Added initial Makefile for building and running project with convenience.
Introduced a Makefile with 'targets' to build the "fluix.exe" executable from source files and a 'run' target to execute the binary, both at once. This came into place now, because the project is growing and it now has multiple files to deal with, because of which, it was getting inconvenient to run long "gcc" commands for multiple files (with different directories) in the terminal and then run the main executable. This, was a nice hack because of which I got to know about something new and learnt a little about Makefiles.